Split HTML chrome renderer tests
