+++ docs/DECISIONS.md
@@ -313,3 +313,4 @@
 - Model and configuration parity gates live in `ParityModelGateTests`, not the broad `ParityGateTests` catch-all. Nike 1.0/Synth branding, TrustedRouter aliases, model catalog normalization, and app config boundaries should stay together so model naming and picker work can evolve without reintroducing Fusion as user-facing copy.
 - Top-bar Disconnect All is a real command, not a disabled placeholder. It shares the active-work stop path with Stop All, stops active MCP server processes, cancels active sends and terminal runs, and detaches the currently selected SSH Remote project context without removing the project from the sidebar. SSH Remote shell/git commands remain noninteractive per-command executions, so disconnecting a remote project clears selection and terminal context rather than closing a persistent SSH socket.
 - Workspace model-picker integration coverage lives in `WorkspaceModelPickerSurfaceIntegrationTests`, not the broad `WorkspaceSurfaceTests` catch-all. Category grouping, model search against workspace state, unknown selected models, recent/favorite ordering, and model badge metadata should stay together while pure DTO compatibility remains in `QuillCodeTopBarSurfaceTests` and pure builder behavior remains in `WorkspaceModelCatalogSurfaceBuilderTests`.
+- HTML chrome renderer coverage lives in `WorkspaceHTMLChromeRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Static HTML smoke coverage for primary regions, sidebar chrome, top-bar overflow, composer markup, context banners, runtime issues, and sidebar date buckets should stay together; tool-card, terminal, browser, secondary-pane, and review HTML coverage can be split into their own focused suites as those areas evolve.
